# Simplify next_weekday and prev_weekday calculus